mission/#17 엔티티 제작 및 서비스 제작

+++ src/main/java/com/example/umc10th/domain/review/dto/ReviewResDTO.java
@@ -2,13 +2,14 @@ package com.example.umc10th.domain.review.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class ReviewResDTO {
     @Builder
     public record CreateReviewDTO(
             Long reviewId,
-            Float score,
+            BigDecimal score,
             String reviewContent,
             LocalDate createdAt
     ) {}
